+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ### Enhancements
 - **General Changes**
   - Update link to Slice
+- **Reports Added**
+  - Editors can create reports that pull data from Slice on a daily basis
+  - Reports can be grouped on pages that users can see under the home tab on the
+    dashboard
 - **Gem Changes**
   - Update to ruby 2.6.3
   - Update to devise 4.6.2